fix(clippy): remove unnecessary u64 cast in create_workspace_volume handler

+++ orchestrator/core/src/presentation/grpc/server.rs
@@ -1093,7 +1093,7 @@ impl AegisRuntime for AegisRuntimeService {
                 volume_name,
                 tenant_id.clone(),
                 storage_class,
-                size_limit_mb as u64,
+                size_limit_mb,
                 ownership,
             )
             .await
